moved the logseq code out, renamed the old index to calendar
